A chat Ui with notes on the second bar
Removed the image analyser. will add in the future. For now you can use the chat UI and chat as long as you can , as it has a n local llm and db. 
future scope.
to achieve fine tuning with emotions from the chat history
to achieve RAG with the chat db to make understand what we actually do with our app in a daily basis and repeat suggestions in the most accepted "okay" way.